requirements.txt
astroid==2.4.0
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
idna==2.9
infi.systray==0.1.12
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
pycodestyle==2.5.0
pylint==2.5.0
pystray==0.15.0
requests==2.23.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
